requirements.txt
aiohttp==3.9.5
aiomcache==0.8.2
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
bitstruct==8.19.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
detools==0.53.0
dulwich==0.20.50
Dumper==1.2.0
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.103.2
fastapi-cache2==0.2.1
feedparser==6.0.11
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==1.34.1
google-auth==2.29.0
google-cloud-pubsub==2.1.0
googleapis-common-protos==1.63.0
grimoirelab-toolkit==1.0.0
grpc-google-iam-v1==0.12.7
grpcio==1.60.0
grpcio-status==1.48.2
h11==0.14.0
heatshrink2==0.13.0
httptools==0.6.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
isodate==0.6.1
libcst==1.3.1
lz4==4.3.3
marshmallow==3.21.1
minio==7.2.5
multidict==6.0.5
neo4j>=5.2.1, ==5.*
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pendulum==3.0.0
perceval==1.0.0
proto-plus==1.23.0
protobuf==3.20.3
pyarrow==15.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.2
pydantic_core==2.14.5
pyelftools==0.30
PyJWT==2.8.0
pymilvus==2.3.7
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.5.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
time-machine==2.14.1
typing_extensions==4.7.1
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
wheel==0.43.0
yarl==1.9.4
zipp==3.15.0
zstandard==0.22.0
